(#31) Apply styling to the Wio terminal display
Organize and beautify the Wio terminal display.

Display the Wifi and MQTT status with colors such that green means
"connected" and red "disconnected".

Add an info and warning banner to display messages such as when readings
are paused for maintenance and when readings are out of bounds.

Add a display interval to the IntervalController.